fix invalid type

+++ src/core/log.c
@@ -133,7 +133,7 @@ void swLog_put(int level, char *content, size_t length)
         break;
     }
 
-    n = sw_snprintf(log_str, SW_LOG_BUFFER_SIZE, "[%.*s %c%d.%d]\t%s\t%.*s\n", l_data_str, date_str, process_flag,
+    n = sw_snprintf(log_str, SW_LOG_BUFFER_SIZE, "[%.*s %c%d.%d]\t%s\t%.*s\n", (int) l_data_str, date_str, process_flag,
             SwooleG.pid, process_id, level_str, (int) length, content);
     if (is_file && flock(SwooleG.log_fd, LOCK_EX) == -1)
     {
